third_party/WebKit/LayoutTests/fast/dom/custom/document-register-reentrant-throwing-constructor.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script src="./resources/document-register-fuzz.js"></script>
</head>
<body>
<div id="container"></div>
<script>
description("Fuzzing document.registerElement() through getters. PASS unless crash.");

setupObjectHooks({
    prototypeGet: function() { throw "Error"; },
    prototypeSet: function(value) { throw "Error"; },
    constructorGet: function() { throw "Error"; },
    constructorSet: function(value) { throw "Error"; }
});

exerciseDocumentRegister();
</script>
</body>
</html>
