third_party/WebKit/LayoutTests/fast/dom/custom/document-register-reentrant-throwing-constructor-expected.txt

CONSOLE MESSAGE: line 53: Error
Fuzzing document.registerElement() through getters. PASS unless crash.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
